requirements.txt
python-dotenv==1.1.0
crewai==0.108.0
crewai-tools==0.38.1
openai==1.70.0
langchain-openai<0.3.0,>=0.2.1
